+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 All notable changes to this project will be documented in this file. Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), with added/changed/deprecated/removed/fixed/security.
 
 ## [unreleased]
+## [0.2.1] - 2023-04-22
 
 ### Fixed
-- 
+- Erros from division by 0
+- Last phrase marked in word
 
 ## [0.2.0] - 2023-02-24
 
@@ -31,7 +33,8 @@ All notable changes to this project will be documented in this file. Based on [K
 ### Added
 - Initial app with words api and flashcard creation
 
-[unreleased]: https://github.com/leafarlins/gokopa/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/leafarlins/gokopa/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/leafarlins/lang/compare/v0.2.0..0.2.1/
 [0.2.0]: https://github.com/leafarlins/lang/compare/v0.1.2..0.2.0/
 [0.1.2]: https://github.com/leafarlins/lang/compare/v0.1.1..0.1.2/
 [0.1.1]: https://github.com/leafarlins/lang/compare/v0.1.0..0.1.1/